Check return codes for error logging
Checks the return code, for 0 (no error) double checks the size downloaded matches the size expected. Saves that error and any other error to a separate error log (to have detailed file information for fixing it) and also adds to the seed-level download log if there were errors or not.