Error handling if doesn't make seed folder
If can't make the folder because the regex to rename doesn't work (AttributeError) or there is a character in the URL that Windows doesn't permit (OSError), have make_seed_folder() raise an error so the script can add the problem to the log and start the next seed. Otherwise, it breaks the script to try to download when the folder wasn't made.