Error log: write immediately
Instead of saving to dictionary and then writing text to log, write to text immediately so the reason for the error can be included. Does cause a problem for the download log that is supposed to know if a seed has a error or not, so will need to add that back in.