Include AIT collection document
ait_collections.py was configuration.py and was not included in the GitHub repository to keep the information private. However, the only information included ended up being AIT collection numbers, which are part of the public URLs, so there is no need to keep this out of GitHub and require UGA users to recreate it for each local machine.

Renamed from configuration.py to ait_collections.py to be more descriptive, added notes to that file to explain the data better, deleted the configuration template, and removed from .gitignore